+++ tests/test_jws.py
@@ -1,6 +1,6 @@
 import pytest
-from superjwt.definitions import DefaultHeadersValidationModel, JOSEHeader
-from superjwt.exceptions import InvalidHeaderError, JWTError
+from superjwt.definitions import JOSEHeader
+from superjwt.exceptions import HeaderValidationError, InvalidHeaderError, JWTError
 from superjwt.jws import JWS
 from superjwt.keys import OctKey
 
@@ -22,21 +22,16 @@ def test_not_reset_jws_instance(
     key = OctKey.import_key(secret_key)
     jws_HS256.encode(
         headers=JOSEHeader(alg="HS256"),
-        payload=claims_fixed_dt,
+        payload=claims_fixed_dt.to_dict(),
         key=key,
-        validation_headers=DefaultHeadersValidationModel,
     )
 
     # not reset JWS instance
     with pytest.raises(JWTError):
-        jws_HS256.decode(
-            token=compact, key=key, validation_headers=DefaultHeadersValidationModel
-        )
+        jws_HS256.decode(token=compact, key=key)
 
     jws_HS256.reset()
-    decoded_claims_after_reset = jws_HS256.decode(
-        token=compact, key=key, validation_headers=DefaultHeadersValidationModel
-    )
+    decoded_claims_after_reset = jws_HS256.decode(token=compact, key=key)
     assert decoded_claims_after_reset.decoded.payload == claims_fixed_dt.to_dict()
 
 
@@ -52,9 +47,7 @@ def test_jws_hmac_decoding(jws_HS256: JWS, claims_fixed_dt, secret_key: str):
 
     key = OctKey.import_key(secret_key)
     decoded_claims = JWTCustomClaims(
-        **jws_HS256.decode(
-            token=compact, key=key, validation_headers=DefaultHeadersValidationModel
-        ).decoded.payload
+        **jws_HS256.decode(token=compact, key=key).decoded.payload
     )
     assert decoded_claims.to_dict() == claims_fixed_dt.to_dict()
 
@@ -73,8 +66,20 @@ def test_wrong_header_algorithm(
     key = OctKey.import_key(secret_key)
     headers = JOSEHeader(alg="HS256")
     headers.alg = "ABCDEF"  # wrong algorithm in header  # type: ignore
+
+    with pytest.raises(InvalidHeaderError):
+        jws_HS256.encode(
+            headers=headers,
+            payload=claims_fixed_dt.to_dict(),
+            key=key,
+        )
+    jws_HS256.reset()
+
     invalid_compact = jws_HS256.encode(
-        headers=headers, payload=claims_fixed_dt, key=key, validation_headers=None
+        headers=headers,
+        payload=claims_fixed_dt.to_dict(),
+        key=key,
+        validation_headers=None,
     ).decode("utf-8")
 
     # not reset JWS instance
@@ -82,24 +87,23 @@ def test_wrong_header_algorithm(
         jws_HS256.decode(
             token=invalid_compact,
             key=key,
-            validation_headers=DefaultHeadersValidationModel,
         )
-
     jws_HS256.reset()
-    with pytest.raises(InvalidHeaderError):
+
+    # header validation error, alg is not a valid algorithm
+    with pytest.raises(HeaderValidationError):
         jws_HS256.decode(
             token=invalid_compact,
             key=key,
-            validation_headers=DefaultHeadersValidationModel,
         )
     jws_HS256.reset()
-    jws_token = jws_HS256.decode(token=invalid_compact, key=key, validation_headers=None)
-    assert jws_token.decoded.headers["alg"] == headers.alg
 
+    # algorithm mismatch error
+    with pytest.raises(InvalidHeaderError):
+        jws_HS256.decode(token=invalid_compact, key=key, validation_headers=None)
     jws_HS256.reset()
+
     decoded_claims = JWTCustomClaims(
-        **jws_HS256.decode(
-            token=compact, key=key, validation_headers=DefaultHeadersValidationModel
-        ).decoded.payload
+        **jws_HS256.decode(token=compact, key=key).decoded.payload
     )
     assert decoded_claims.to_dict() == claims_fixed_dt.to_dict()